# servers/node/src/docs/openapi.yaml
openapi: "3.0.0"
info:
  version: 12.0.0
  title: Jeera Project Manager
  license:
    name: MIT
  contact:
    name: Antoine Cheron
    email: cheron.antoine@gmail.com
paths:
...
  /projects:
    get:
      summary: List the projects that the connected user can see
      operationId: listProjects
      security:
        - bearerToken: []
      tags:
        - projects
      parameters:
        - name: offset
          in: query
          description: The number of items to skip before starting to collect the result set
          required: false
          schema:
            type: integer
            format: int8
            minimum: 0
            default: 0
        - name: limit
          in: query
          description: Number of elements to get
          required: false
          schema:
            type: integer
            format: int8
            minimum: 1
            maximum: 40
            default: 5
      responses:
        '200':
          description: 'Success'
          headers:
            X-Next:
              description: Link to next page
              required: false
              schema:
                type: string
                format: url
            X-Last:
              description: Link to last page
              required: false
              schema:
                type: string
                format: url
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectsList'
          links:
            createProject:
              operationId: createProject
            setTaskStatusesFlow:
              operationId: setProjectTaskStatusFlow
            setDetails:
              operationId: setProjectDetails
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      summary: Creates a project
      operationId: createProject
      tags:
        - projects
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          $ref: '#/components/responses/projectDetails'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security: 
        - bearerToken: []
  /project/{projectId}/task-status-flow:
    parameters: 
      - $ref: '#/components/parameters/projectId'
    post:
      summary: Set the tasks status flow of the project
      operationId: setProjectTaskStatusFlow
      tags:
        - projects
      security: 
        - bearerToken: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                taskStatuses:
                  type: array
                  minimum: 1
                  items:
                    $ref: '#/components/schemas/TaskStatus'
                taskStatusTransitions:
                  type: array
                  minimum: 1
                  items:
                    $ref: '#/components/schemas/TaskStatusTransition'
              required:
                - taskStatuses
                - taskStatusTransitions

      responses:
        '200':
          $ref: '#/components/responses/projectDetails'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /project/{projectId}/details:
    parameters: 
      - $ref: '#/components/parameters/projectId'
    post:
      summary: Set the project details
      operationId: setProjectDetails
      tags:
        - projects
      security: 
        - bearerToken: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                description:
                  type: string
                collaborators:
                  type: array
                  items:
                    type: string
                    description: The name of the users to add as collaborators of the project
              required:
                - description
      responses:
        '200':
          $ref: '#/components/responses/projectDetails'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /project/{projectId}:
    parameters: 
        - $ref: '#/components/parameters/projectId'
    get:
      summary: Get the detail of a project
      operationId: getProjectDetails
      tags:
        - projects
      responses:
        '200':
          $ref: '#/components/responses/projectDetails'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security: 
        - bearerToken: []
    delete:
      summary: Delete a project (only if archived and the user is an admin)
      operationId: deleteProject
      tags:
        - projects
      responses:
        '204':
          description: Success
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security: 
        - bearerToken: []
...
  /project/{projectId}/star:
    post:
      description: Reverse starred status of a project
      operationId: starProject
      parameters: 
        - $ref: '#/components/parameters/projectId'
      tags:
        - projects
      responses:
        '204':
          description: Success
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
        - bearerToken: []
  /project/{projectId}/addCollaborator:
    post:
      description: Invite users to the project
      operationId: inviteUsersToProject
      parameters: 
        - $ref: '#/components/parameters/projectId'
      tags:
        - projects
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InviteUsersToProjectRequest'
      responses:
        '204':
          description: Success
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security: 
        - bearerToken: []
  /project/{projectId}/archive:
    post:
      description: Archive a project
      operationId: switchProjectArchiveState
      tags:
        - projects
      parameters: 
        - $ref: '#/components/parameters/projectId'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  isArchived:
                    type: boolean
                required:
                  - isArchived
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security: 
        - bearerToken: []
  /tasks:
    parameters: 
        - $ref: '#/components/parameters/projectId'
    get:
      summary: List user tasks
      operationId: listTasks
      tags:
        - tasks
      parameters:
        - $ref: '#/components/parameters/queryProjectId'
        - name: createdBefore
          in: query
          description: Minimum date of tasks to list
          required: false
          schema:
            type: string
            format: date
        - name: offset
          in: query
          description: The number of items to skip before starting to collect the result set
          required: false
          schema:
            type: integer
            format: int8
            minimum: 0
            default: 0
        - name: limit
          in: query
          description: Number of elements to get
          required: false
          schema:
            type: integer
            format: int8
            minimum: 1
            maximum: 40
            default: 10
      responses:
        '200':
          description: Success - list of tasks
          headers:
            Link:
              description: 'Contextually relevant links - provided links all have a rel attribute as recommended here: https://www.w3.org/wiki/LinkHeader'
              schema:
                type: string
                format: link-http-header
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TasksList'
          links:
            createTechnicalStory:
              operationId: createTechnicalStory
              parameters:
                parentProjectId: "$request.query.queryProjectId"
            createUserStory:
              operationId: createUserStory
              parameters:
                parentProjectId: "$request.query.queryProjectId"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security: 
        - bearerToken: []
...
components:
  securitySchemes:
    bearerToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    projectDetails:
      description: Project details
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Project'
      links:
        inviteUser:
          operationId: inviteUsersToProject
          parameters:
            projectId: '$response.body#/projectId'
        createTechnicalStory:
          operationId: createTechnicalStory
          parameters:
            parentProjectId: '$response.body#/id'
        createUserStory:
          operationId: createUserStory
          parameters:
            parentProjectId: '$response.body#/id'
        listTasks:
          operationId: listTasks
          parameters:
            queryProjectId: '$response.body#/id'
        reverseArchivedState:
          operationId: switchProjectArchiveState
          parameters:
            projectId: '$response.body#/id'
        delete:
          operationId: deleteProject
          parameters:
            projectId: '$response.body#/id'
        star:
          operationId: starProject
          parameters:
            projectId: '$response.body#/id'
        analytics:
          operationId: getAnalytics
          parameters:
            resourceId: '$response.body#/id'
        setTaskStatusesFlow:
          operationId: setProjectTaskStatusFlow
          parameters:
            projectId: '$response.body#/id'
        setDetails:
          operationId: setProjectDetails
          parameters:
            projectId: '$response.body#/id'
  parameters:
...
    projectId:
      name: projectId
      in: path
      description: Id of the project to retrieve details for
      required: true
      schema:
        type: string
        format: url
...
    queryProjectId:
      name: queryProjectId
      in: query
      description: Id of the project to retrieve details for
      required: true
      schema:
        type: string
        format: url
  schemas:
...
    CreateProjectRequest:
      type: object
      properties:
        name:
          type: string
      required: [ name ]
    ProjectsList:
      type: object
      properties:
        projects:
          type: array
          items:
            $ref: '#/components/schemas/Project'
    Project:
      type: object
      properties:
        name:
          type: string
        id:
          type: string
        isArchived:
          type: boolean
        collaborators:
          type: array
          items:
            type: string
            description: username
        description:
          type: string
        availableTaskStatuses:
          type: array
          items:
            $ref: '#/components/schemas/TaskStatus'
        taskStatusTransitions:
          type: array
          items:
            $ref: '#/components/schemas/TaskStatusTransition'
        nextCreationStep:
          type: string
          enum: 
            - INITIALIZATION
            - TASK_FLOW_SETUP
            - INFORMATION_SETUP
            - CREATION_COMPLETED
      required: [ name, id, isArchived, collaborators ]
    TasksList:
      type: object
      properties:
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/Task'
...
    TechnicalStory:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        parentProjectId:
          type: string
        details:
          $ref: '#/components/schemas/TaskDetails'
        tags:
          type: array
          items:
            type: string
        priority:
          $ref: '#/components/schemas/Priority'
        isArchived:
          type: boolean       
      required: [ id, title, parentProjectId, details, isArchived ]
    TaskDetails:
      type: object
      properties:
        description:
          type: string
        assignee:
          type: string
        status:
          type: string
...
    UserStory:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        points:
          type: number
        parentProjectId:
          type: string
        details:
          $ref: '#/components/schemas/TaskDetails'
        tags:
          type: array
          items:
            type: string
        priority:
          $ref: '#/components/schemas/Priority'
        isArchived:
          type: boolean
      required: [ id, title, points, parentProjectId, details, isArchived ]
    Task:
      oneOf:
        - $ref: '#/components/schemas/TechnicalStory'
        - $ref: '#/components/schemas/UserStory'
    TaskStatus:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
    TaskStatusTransition:
      type: object
      properties:
        from:
          type: string
          description: Either a comma-delimited string of TaskStatus ids or '*' to select them all
        to:
          type: string
          description: Either a comma-delimited string of TaskStatus ids or '*' to select them all
    Priority:
      type: string
      enum:
        - blocking
        - critical
        - important
        - high
        - medium
        - low
        - simple
...
    Error:
      type: object
      properties:
        code:
          type: string
        description:
          type: string
    InviteUsersToProjectRequest:
      type: object
      properties:
        users:
          type: array
          items:
            type: string
            description: username
          minItems: 1
          maxItems: 5
...